fix: interrupt number var/param size
interrupt number is uint8_t, not uint16_t

+++ cc/lib/time_timer.64.c
@@ -16,7 +16,7 @@ void time_timer_reset_tick_count() {
 	time_timer_tick_count = 0;
 }
 
-void time_timer_pit_isr(interrupt_frame_t* frame, uint16_t intnum){
+void time_timer_pit_isr(interrupt_frame_t* frame, uint8_t intnum){
 	UNUSED(frame);
 	UNUSED(intnum);
 
@@ -38,7 +38,7 @@ void time_timer_pit_sleep(uint64_t usecs) {
 	while(time_timer_tick_count <= usecs);
 }
 
-void time_timer_apic_isr(interrupt_frame_t* frame, uint16_t intnum) {
+void time_timer_apic_isr(interrupt_frame_t* frame, uint8_t intnum) {
 	UNUSED(frame);
 	UNUSED(intnum);
